Vote and delete opinions

+++ app/Http/Controllers/OpinionController.php
@@ -6,6 +6,7 @@ use App\Http\Requests\OpinionCreateRequest;
 use Illuminate\Support\Facades\Auth;
 use App\Models\Opinion;
 use App\Models\Practice;
+use App\Helpers\Vote;
 
 class OpinionController extends Controller
 {
@@ -36,11 +37,44 @@ class OpinionController extends Controller
     public function delete(int $id)
     {
         $opinion = Opinion::find($id);
-        $opinion->delete();
+        if (!Auth::user()->wrote($opinion)) {
+            session()->flash('message', 'Vous ne pouvez pas suprimmer des opinions autres que les vôtres!');
+            return redirect()->back();
+        } else {
+            $opinion->delete();
+            session()->flash('message', 'Votre opinion a été suprimmée');
+            return redirect()->back()->withInput();
+        }
+    }
 
-        session()->flash('message', 'Opinion successfully deleted');
+    private function vote(int $opinion_id, int $vote)
+    {
+        $opinion = Opinion::findOrFail($opinion_id);
 
+        if (Auth::user()->wrote($opinion)) {
+            session()->flash('message', 'Vous ne pouvez pas vote votre opinion !');
+            return redirect()->back()->withInput();
+        }
+
+        if ($vote === Vote::UPVOTE) {
+            $opinion->upvote();
+        } else {
+            $opinion->downvote();
+        }
+
+        session()->flash('message', "Vous avez voté l'opinion avec succès !");
         return redirect()->back()->withInput();
+        
+    }
+
+    public function upvote(int $opinion_id)
+    {
+        return $this->vote($opinion_id, Vote::UPVOTE);
+    }
+
+    public function downvote(int $opinion_id)
+    {
+        return $this->vote($opinion_id, Vote::DOWNVOTE);
     }
 
 }
